fix(async): cast the token feedback to the staged input dtype
The feedback scatter claimed to take the sampling op's dtype whatever it is,
but only its identity branch actually did: copy_ converts, index_put_ demands
equal dtypes. An all-greedy batch reaches `rbln::argmax`, which still declares
int64, so any step whose batch order changed died on a dtype mismatch against
the int32 staged buffer.

vLLM-native path.

+++ vllm_rbln/v1/worker/rbln_model_runner.py
@@ -1679,7 +1679,12 @@ class RBLNModelRunner(KVConnectorModelRunnerMixin):
                     host.copy_(prev_sampled[:n_fb, 0])
                     staged_model_inputs.input_ids[:n_fb, 0].copy_(host)
                 elif rows:
-                    staged_model_inputs.input_ids[dsts, 0] = prev_sampled[rows, 0]
+                    # index_put_ demands equal dtypes, unlike the copy_ above.
+                    # The sampling op's width is not fixed -- `rbln::argmax`
+                    # still returns int64 -- so cast to the staged buffer's.
+                    staged_model_inputs.input_ids[dsts, 0] = prev_sampled[rows, 0].to(
+                        staged_model_inputs.input_ids.dtype
+                    )
 
         # Run the model.
         # With spec decode, defer connector finalization (wait_for_save + clear
